[ALPS07439323] wifi: Fix coverity issue
1. [Coverity][ALPS]CERT-C Characters and Strings: 10645364

MTK-Commit-Id: ebc64e60f3967b8cbb3a118506cab873c3a3bfea

CR-Id: ALPS07439323
Feature: others

+++ common/wlan_lib.c
@@ -4884,8 +4884,8 @@ uint32_t wlanGetMiniTxPower(IN struct ADAPTER *prAdapter,
 		return WLAN_STATUS_NOT_ACCEPTED;
 	}
 
-	startOfs = arRange[bandIdx][ePhyMode].startOfs;
-	endOfs = arRange[bandIdx][ePhyMode].endOfs;
+	startOfs = arRange[(uint8_t)bandIdx][(uint8_t)ePhyMode].startOfs;
+	endOfs = arRange[(uint8_t)bandIdx][(uint8_t)ePhyMode].endOfs;
 
 
 	/*NVRAM start addreess :0*/
